requirements.txt
ollama
GitPython
openai
markdown_pdf
pyfiglet